+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Major
 * Introduced package with functions to demultiplex fastq files, align either to a reference or denovo, and genotype resulting alignments using ANGSD.
+* Introduced a `shiny` GUI with support for all of `alignR`'s major functions.
 
 ## Documentation
 * Functions are all documented with usage, parameters, and return. Need references and examples (alongside example data).